Be counts

+++ be/src/main/java/com/brodeckyondrej/SignUp/business/dto/subject/SubjectGetListDto.java
@@ -7,10 +7,10 @@ import java.util.UUID;
 
 @Getter
 public class SubjectGetListDto extends NamedDtoWithId {
-    private final int numberOfStudents;
-    private final int numberOfCategories;
+    private final long numberOfStudents;
+    private final long numberOfCategories;
 
-    public SubjectGetListDto (UUID id, String name, int numberOfStudents, int numberOfCategories){
+    public SubjectGetListDto (UUID id, String name, long numberOfStudents, long numberOfCategories){
         super(id, name);
         this.numberOfStudents = numberOfStudents;
         this.numberOfCategories = numberOfCategories;
